etnaviv: implement resource creation with modifier
This implements resource creation with modifiers for etnaviv. A suitable
modifier for the render target is chosen from the list of available
modifers.

If we are able to use tiling, the imported resource will used as the
base resource directly, so we don't need to change the logic at other
places in the driver.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>