etnaviv: turn scanout surface into one type of external surface
Scanout is just a specialization of an external surface, which might not
be compatible with the sampler or pixel engines. Handle it accordingly.

This allows us to import other types of external surfaces with
incompatible tiling. Use this to implement dma-buf import/export with the
correct tiling information and resolve steps.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>