DRI: add image usage flags to optimize buffer allocation
For now all DRI images implicitly assume both render and texture usage,
as well as no (or slow) CPU access. Creating DRI images is the only way
for GBM to allocate buffers. With gbm_gralloc the usage of the buffer
might be something wildly different from the current default, so DRI
image need to grow some flags to describe those.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>