etnaviv: keep cached shadow for staging resource
For resources that need fast CPU access it is beneficial to keep a shadow
copy in cached memory. We don't switch the BO to cached, as benchmarks
have shown that the needed cache maintenance on cached BOs is marginally
slower than the current unoptimized tile upload. With some more
optimizations to the upload path, the performance advantage of the copy
will likely be even bigger.

Currently this is only done for sampler resources, so we don't have to deal
with buffer content changed by the GPU, which would require even more seqno
tracking.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>

Conflicts:
	src/gallium/drivers/etnaviv/etnaviv_transfer.c

Fixed conflict due to the changes introduced in 17e8c240231 with the
calls to etna_copy_resource_box().

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>